# [Numpy] FFI: tril_indices