@XIAO-XIA 2020-06-15
@mxnet-bot run ci [unix-cpu, unix-gpu]