@XIAO-XIA 2020-06-16
> @XIAO-XIA CI fails in a test related to the Op you're modifying. It suggests there's a bug in your PR

Thank you very much! I'm trying to fix the bug.