@XIAO-XIA 2020-06-17
@mxnet-bot run ci [unix-cpu, unix-gpu]