@XIAO-XIA 2020-06-20
@mxnet-bot run ci [unix-cpu]