@XIAO-XIA 2020-06-21
@mxnet-bot run ci [unix-cpu]